Only find boot/data_part if we're building hybris-boot
hybris-boot has a nasty tendency to throw off the entire build if the
fstab and fixup-mountpoints file aren't correct for the current device.
This is fine when hybris-boot is the only possible bootimage in the
tree, but I expect that we'll have it alongside halium-boot for a little
while longer. By putting the partition evaluation and subsequent build
error inside the local module's main rule, we keep them from failing the
build unless we're actually building hybris-boot.

Change-Id: If3ed791d0ba3c11a84d11cef0ea8ef00ab272e9a